+++ packages/datatrak-web/src/database/index.ts
@@ -1,3 +1,4 @@
 export * from './entity';
 export * from './task';
 export * from './user';
+export * from './clearDatabase';
